feat: 앱 아이콘 및 스플래시 화면 (#5)
