+++ GWNU Club/Views/SplashView.swift
@@ -1,0 +1,20 @@
+//
+//  SplashView.swift
+//  GWNU Club
+//
+//  Created by 박세민 on 10/14/24.
+//
+
+import SwiftUI
+
+struct SplashView: View {
+    var body: some View {
+        Image("Splash")
+            .resizable()
+            .scaledToFit()
+    }
+}
+
+#Preview {
+    SplashView()
+}
